+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.0.1
 
--   Upgrade to eslint 5 (5.9.0)
--   Add @babel/plugin-proposal-export-default-from
+-   Upgrade to `eslint@5.9.0`
+-   Add `@babel/plugin-proposal-export-default-fro`m
+-   Add `create` command